+++ packages/teams-js/src/internal/videoPeformanceMonitor.ts
